Tempth 1 for tilted
One possibility of introducing the tilted geometry is create a new
branch from TrackerLayerBuilder and choose, instead of Ladder, Rings.
This is the approach that has been implemented in this commit and works
fine: some changes are needed into the xml file produced to recognize
the rings inside the barrel.
Two main issues raised:
1) the DetId is still fine?
2) how to do the ordering in the single layer of the barrel?