TrackerTopology isDoubleSide
The method compose() into the SiStripDetIds is added in order to take
into account the possibility of the presence of PhaseII stacks.
In the TkTopology, the definition of the double sided detector is given
depending on the geometry (upgraded or not).
If the bool upgrade is set True, the OT geometry is consider to be made
by stacks, otherwise the definition is kept as before (glued + specific layers).
The print method into the cc. is updated as should be.