Introducing StackGeomDet
This class has a similar function to the GluedGeomDet one. It is
composed by the Lower and the Upper det and its DetId is defined with
the stack function defined in SiStripDetId. This last point is done in
contrast with the used of the SubDet into the Glued, because the
information into the SiStripDetId are also used by the Topology, so it
seems to me more consistent.