Adding stacks in the MeasurementTracker
The main change has been done into the process of adding GeomDet.
Now every subdetector containing modules calls the function addDets which can decide the
different nature of the GeomDet - with the information coming from
isTrackerPixel(). In this way, both pixel and strip subdets have the
possibility to contain Glued or Stack.
The stacks are added exactly in the same way that has been done for the
GluedGeomDet using the functions addStackDet and initStackDet.
A trivial check is added at the end of the filling process in order to
compare the number of modules in the TrackerGeometry and in the
DetContainer.