Removing SiStripDetId dependences
- In TrackerGeomBuilderFromGeometricDet the dependence of the SiStripDetId
has been removed using the GeomType information inside to the GeomDet.
Using the name of the module is possible to check directly from the xml
file if the module is Stereo or Lower and then create the Glued or the
Stack respectively.
- In Glued- and StackGeomDet the dependence of the SiStripDetId has been removedpassing directly the detId of the composed module.

Moreover:
Add in the tkTopology the same of the previous commit
(#63d201d209272b16ef0b62ecb9fc2b44dafcb82f) for RPhi.