Attempt 2 for tilted
The second possibility is to build another Layer
(Phase2OTBarrelTiltedLayer) which takes into the possibility to have
both rings and rods. The ordering is done introducing a two bits in the
DetId correspondent to:
1 = rings-
2 = rods
3 = rings+
and ordering them by z and phi.
This is the approach that has been implemented in this commit and works
fine: some changes are needed into the xml file produced to recognize
the rings inside the barrel.
I have also check with the Navigation school that the stucture of the navigable
layers is ok (a bit changed by the one without tilted modules).

Question:
1) This kind of layer is just used for the Layer 1, 2, 3 ... should be
extended?