Bool variable useStacks
In order to maintain both the "official" reconstruction and the one with
VH in phase2 tracking, I have introduced a new bool variable in
TrackerRecoGeometryESProducer that allow to create brothers or not
(respectively in the two different reco alternatives).
Still to test if the "off" reco is already working or need some
adjustment more.